Remove Polymer integration & move app to demo path
