+++ bower.json
@@ -12,20 +12,6 @@
     "**/.*",
     "public/components"
   ],
-  "dependencies": {
-    "app-layout": "PolymerElements/app-layout#^0.10.0",
-    "iron-icons": "PolymerElements/iron-icons#^1.1.3",
-    "iron-icon": "PolymerElements/iron-icon#^1.0.9",
-    "paper-card": "PolymerElements/paper-card#^1.1.1",
-    "paper-header-panel": "PolymerElements/paper-header-panel#^1.1.6",
-    "paper-item": "PolymerElements/paper-item#^1.2.1",
-    "paper-menu": "PolymerElements/paper-menu#^1.2.2",
-    "paper-styles": "PolymerElements/paper-styles#^1.1.4",
-    "polymer": "Polymer/polymer#^1.4.0",
-    "paper-icon-button": "PolymerElements/paper-icon-button#^1.1.2"
-  },
-  "resolutions": {
-    "polymer": "^1.6.0",
-    "paper-styles": "^1.1.4"
-  }
+  "dependencies": {},
+  "resolutions": {}
 }
