--- client/imports/app/app.component.html
@@ -1,16 +1,0 @@
-<app-drawer-layout force-narrow fullbleed>
-    <app-drawer #navigation swipe-open [opened]="toolbar.opened" (app-drawer-transitioned)="toolbar.opened = navigation.opened">
-        <mantra-navigation></mantra-navigation>
-    </app-drawer>
-
-    <!-- FIXME: move content out of app-header-layout, see here: https://github.com/PolymerElements/app-layout/issues/155 -->
-    <app-header-layout>
-        <app-header shadow>
-            <mantra-toolbar #toolbar></mantra-toolbar>
-        </app-header>
-    </app-header-layout>
-
-    <div class="content" fullbleed>
-        <router-outlet></router-outlet>
-    </div>
-</app-drawer-layout>
