--- client/imports/app/toolbar/toolbar.component.html
@@ -1,4 +1,0 @@
-<app-toolbar>
-    <paper-icon-button icon="{{opened ? 'icons:arrow-back' : 'icons:menu'}}" (click)="toggle()" drawer-toggle></paper-icon-button>
-    <div main-title>{{(toolbar$ | async).title}}</div>
-</app-toolbar>
