+++ client/imports/demo/+home/home.component.html
@@ -1,3 +1,3 @@
-<paper-card>
+<div id="home">
     <h2>{{(home$ | async).message}}</h2>
-</paper-card>
+</div>
