+++ client/imports/demo/+home/home.routes.ts
@@ -1,5 +1,5 @@
 import { Routes } from '@angular/router';
-import { HomeComponent } from '/client/imports/+home/home.component';
+import { HomeComponent } from '/client/imports/demo/+home/home.component';
 
 export const homeRoutes: Routes = [
     { path: '', component: HomeComponent }
