+++ client/imports/demo/+home/home.service.ts
@@ -1,8 +1,8 @@
 import { Store } from '@ngrx/store';
-import { AppState } from '/client/imports/app/app.state';
-import { HomeActions } from '/client/imports/+home/home.actions';
+import { AppState } from '/client/imports/demo/app/app.state';
+import { HomeActions } from '/client/imports/demo/+home/home.actions';
 import { Observable } from 'rxjs';
-import { HomeState } from '/client/imports/+home/home.state';
+import { HomeState } from '/client/imports/demo/+home/home.state';
 import { Injectable } from '@angular/core';
 
 @Injectable()
