+++ client/imports/demo/app/app.reducer.ts
@@ -1,5 +1,5 @@
 import { Action, ActionReducer } from '@ngrx/store';
-import { AppState } from '/client/imports/app/app.state';
+import { AppState } from '/client/imports/demo/app/app.state';
 import { initialState as routerInitialState } from '@ngrx/router-store';
 
 const initialState: AppState = <AppState> {
