+++ client/imports/demo/app/app.routes.ts
@@ -1,5 +1,5 @@
 import { Routes } from '@angular/router';
-import HomeModule from '/client/imports/+home/home.module';
+import HomeModule from '/client/imports/demo/+home/home.module';
 
 
 export const appRoutes: Routes = [
@@ -7,7 +7,7 @@ export const appRoutes: Routes = [
     {
         path: 'home',
         loadChildren: (() => new Promise((resolve, reject) => {
-            return resolve(require('/client/imports/+home/home.module').default)
+            return resolve(require('/client/imports/demo/+home/home.module').default)
         })),
         data: {
             reducers: HomeModule.reducers(),
