+++ client/imports/demo/app/app.state.ts
@@ -1,6 +1,6 @@
 import { ToolbarState, SidebarStatus } from './toolbar/toolbar.state';
 import { NavigationSidebarState } from './navigation-sidebar/navigation-sidebar.state';
-import { HomeState } from '/client/imports/+home/home.state';
+import { HomeState } from '/client/imports/demo/+home/home.state';
 import { RouterState } from '@ngrx/router-store';
 
 export interface AppState {
