+++ client/imports/demo/app/navigation-sidebar/menu-item/menu-item.component.html
@@ -1,5 +1,4 @@
-<paper-icon-item>
-    <iron-icon icon="{{_icon}}" item-icon></iron-icon>
+<div class="menu item">
     <a href="{{_link}}">{{_name}}</a>
     <!--<span md-line class="secondary" *ngIf="_description">{{_description}}</span>-->
-</paper-icon-item>
+</div>
