+++ client/imports/demo/app/navigation-sidebar/navigation-sidebar.actions.ts
