+++ client/imports/demo/app/navigation-sidebar/navigation-sidebar.component.html
@@ -1,4 +1,4 @@
-<paper-menu>
+<div class="nav menu">
     <mantra-menu-item *ngFor="let item of (navigation$ | async).menu.items" [item]="item">
     </mantra-menu-item>
-</paper-menu>
+</div>
