+++ client/imports/demo/app/navigation-sidebar/navigation-sidebar.component.ts
