+++ client/imports/demo/app/navigation-sidebar/navigation-sidebar.effects.ts
