+++ client/imports/demo/app/navigation-sidebar/navigation-sidebar.reducer.ts
