+++ client/imports/demo/app/toolbar/toolbar.component.ts
