+++ client/imports/demo/app/toolbar/toolbar.reducer.ts
