+++ client/main.html
@@ -1,6 +1,5 @@
 <head>
     <base href="/">
-    <link rel="import" href="polymer.html">
 </head>
 <body>
     <mantra-app></mantra-app>
