+++ package.json
@@ -24,7 +24,6 @@
     "@ngrx/store": "^2.1.0",
     "@ngrx/store-devtools": "^3.2.2",
     "@ngrx/store-log-monitor": "^3.0.0",
-    "@vaadin/angular2-polymer": "^1.0.0-beta5",
     "angular2-meteor": "^0.7.1",
     "angular2-meteor-polyfills": "^0.1.1",
     "babel-runtime": "^6.18.0",
