--- public/polymer.html
@@ -1,45 +1,0 @@
-<!-- Polymer - Base -->
-
-<script>
-    window.Polymer = {
-        dom: 'shadow',
-        lazyRegister: false
-    };
-</script>
-
-<link rel="import" href="components/polymer/polymer.html">
-
-<!-- Polymer - App Layout -->
-
-<link rel="import" href="components/app-layout/app-layout.html">
-
-<!-- Polymer - Iron -->
-
-<!--<link rel="import" href="components/iron-autogrow-textarea/iron-autogrow-textarea.html">-->
-<link rel="import" href="components/iron-icon/iron-icon.html">
-<link rel="import" href="components/iron-icons/iron-icons.html">
-<!--<link rel="import" href="components/iron-component-page/iron-component-page.html">-->
-<link rel="import" href="components/iron-flex-layout/iron-flex-layout-classes.html">
-<!--<link rel="import" href="components/iron-form/iron-form.html">-->
-<!--<link rel="import" href="components/iron-overlay-behavior/iron-overlay-behavior.html">-->
-
-<!-- Polymer - Paper -->
-
-<link rel="import" href="components/paper-card/paper-card.html">
-<!--<link rel="import" href="components/paper-dialog/paper-dialog.html">-->
-<!--<link rel="import" href="components/paper-drawer-panel/paper-drawer-panel.html">-->
-<!--<link rel="import" href="components/paper-dropdown-menu/paper-dropdown-menu.html">-->
-<link rel="import" href="components/paper-header-panel/paper-header-panel.html">
-<link rel="import" href="components/paper-icon-button/paper-icon-button.html">
-<!--<link rel="import" href="components/paper-input/paper-input.html">-->
-<!--<link rel="import" href="components/paper-input/paper-textarea.html">-->
-<link rel="import" href="components/paper-item/paper-icon-item.html">
-<link rel="import" href="components/paper-item/paper-item.html">
-<link rel="import" href="components/paper-menu/paper-menu.html">
-<!--<link rel="import" href="components/paper-progress/paper-progress.html">-->
-<link rel="import" href="components/paper-styles/paper-styles.html">
-<!--<link rel="import" href="components/paper-toolbar/paper-toolbar.html">-->
-
-<!-- Paper Datatable by https://github.com/David-Mulder/paper-datatable -->
-
-<!--<link rel="import" href="vendor/paper-datatable/paper-datatable-card.html">-->
